Split Cashu transfer operations
