+++ tests/verify-modules.js
@@ -98,6 +98,7 @@
     '/js/onboarding-view.js',
     '/js/data-merge-lab-entries.js',
     '/js/data-view-controls.js',
+    '/js/cashu-wallet-transfers.js',
     '/js/pii-review.js',
     '/js/marker-detail-modal.js',
     '/js/marker-detail-modal-impl.js',
